[larger-key-info] Extract as a generic number picker component
