+++ app/src/main/java/com/litus_animae/refitted/compose/exercise/RepsDisplay.kt
@@ -5,15 +5,12 @@ import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.BoxWithConstraints
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.pager.HorizontalPager
-import androidx.compose.foundation.pager.PageSize
 import androidx.compose.foundation.pager.rememberPagerState
 import androidx.compose.material.Card
 import androidx.compose.material.Divider
@@ -26,14 +23,7 @@ import androidx.compose.runtime.mutableStateListOf
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
-import androidx.compose.ui.BiasAlignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.drawWithContent
-import androidx.compose.ui.graphics.BlendMode
-import androidx.compose.ui.graphics.Brush
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.CompositingStrategy
-import androidx.compose.ui.graphics.graphicsLayer
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.tooling.preview.PreviewParameter
@@ -41,6 +31,7 @@ import androidx.compose.ui.unit.dp
 import com.litus_animae.refitted.R
 import com.litus_animae.refitted.compose.state.ExerciseSetWithRecord
 import com.litus_animae.refitted.compose.state.Repetitions
+import com.litus_animae.refitted.compose.util.NumberPicker
 import com.litus_animae.refitted.models.ExerciseSet
 import com.litus_animae.refitted.models.Record
 import kotlinx.coroutines.flow.emptyFlow
@@ -79,6 +70,8 @@ fun RepsDisplay(
         exerciseSet.isToFailure -> "${setWithRecord.reps} (MAX)"
         else -> "${setWithRecord.reps}"
       }
+      
+      val typography = MaterialTheme.typography.h3
 
       val currentRepsValue by reps.value
       val pagerState = rememberPagerState(currentRepsValue)
@@ -88,58 +81,8 @@ fun RepsDisplay(
           .fillMaxWidth()
       ) {
         val pageWidth = 80.dp
-        HorizontalPager(
-          pageCount = 101,
-          Modifier.width(maxWidth),
-          contentPadding = PaddingValues(horizontal = (maxWidth - pageWidth) / 2),
-          state = pagerState,
-          pageSize = PageSize.Fixed(pageWidth)
-        ) {
-          val pageOffset = (
-            (pagerState.currentPage - it) + pagerState
-              .currentPageOffsetFraction
-            )
-          Box(
-            Modifier
-              .fillMaxSize()
-              .graphicsLayer(compositingStrategy = CompositingStrategy.Offscreen)
-              .drawWithContent {
-                drawContent()
-                if (pagerState.currentPage == it) {
-                  // noop
-                } else if (pageOffset > 0) {
-                  // fade in from left
-                  // first to left has value 1.0 => 0.8
-                  val gradientStart = (pageOffset - 0.2f).coerceIn(0f, 1f)
-                  val gradientEnd = pageOffset.coerceIn(0f, 1f)
-                  drawRect(
-                    brush = Brush.horizontalGradient(
-                      gradientStart to Color.Transparent,
-                      gradientEnd to Color.Red),
-                    blendMode = BlendMode.DstIn
-                  )
-                } else if (pageOffset < 0) {
-                  // fade out to right
-                  // first to right has value -1.0 => 0/0.2
-                  val gradientStart = (pageOffset + 1f).coerceIn(0f, 1f)
-                  val gradientEnd = (pageOffset + 1.2f).coerceIn(0f, 1f)
-                  drawRect(
-                    brush = Brush.horizontalGradient(
-                      gradientStart to Color.Red,
-                      gradientEnd to Color.Transparent),
-                    blendMode = BlendMode.DstIn
-                  )
-                }
-              },
-            contentAlignment = BiasAlignment(pageOffset.coerceIn(-1f, 1f), 1f)
-          ) {
-            // FIXME 100 overflows
-            Text(
-              it.toString(),
-              style = MaterialTheme.typography.h3
-            )
-          }
-        }
+        val pageCount = 101
+        NumberPicker(pageCount, pageWidth, pagerState, typography)
       }
       val lineColor = contentColorFor(MaterialTheme.colors.surface)
       Divider(
@@ -154,7 +97,8 @@ fun RepsDisplay(
           horizontalAlignment = Alignment.CenterHorizontally
         ) {
           // FIXME when "seconds" it doesn't fit
-          Text(targetRepsTextContent, style = MaterialTheme.typography.h3)
+          // TODO probably use java period for localization
+          Text(targetRepsTextContent, style = typography)
         }
       }
